Fixes for Events, Locations, & Tags Routes

+++ services/hexathons/src/routes/location.ts
@@ -1,7 +1,8 @@
 import { asyncHandler, BadRequestError, checkAbility } from "@api/common";
 import express from "express";
 
-import { LocationModel, Location } from "../models/location";
+import { EventModel } from "../models/event";
+import { LocationModel } from "../models/location";
 
 export const locationRoutes = express.Router();
 
@@ -11,6 +12,14 @@ locationRoutes.route("/").get(
     const locations = await LocationModel.accessibleBy(req.ability);
 
     return res.send(locations);
+  })
+);
+
+locationRoutes.route("/:id").get(
+  checkAbility("read", "Location"),
+  asyncHandler(async (req, res) => {
+    const location = await LocationModel.findById(req.params.id).accessibleBy(req.ability);
+    return res.send(location);
   })
 );
 
@@ -40,8 +49,8 @@ locationRoutes.route("/:id").put(
       name: req.body.name,
     });
 
-    if (existingLocation) {
-      throw new BadRequestError(`Location with name ${  req.body.name  } already exists`);
+    if (existingLocation?.id !== req.params.id) {
+      throw new BadRequestError(`Location with name ${req.body.name} already exists`);
     }
 
     const location = await LocationModel.findByIdAndUpdate(
@@ -61,6 +70,12 @@ locationRoutes.route("/:id").put(
 locationRoutes.route("/:id").delete(
   checkAbility("delete", "Location"),
   asyncHandler(async (req, res) => {
+    const event = await EventModel.findOne({
+      location: req.params.id,
+    });
+    if (event) {
+      throw new BadRequestError("Event with location still exists. Location cannot be deleted.");
+    }
     await LocationModel.findByIdAndDelete(req.params.id);
     return res.sendStatus(204);
   })
